Adds About command
